+++ scripts/troll_exp.R
@@ -18,14 +18,12 @@ test <- snakemake@params$test
 test_years <- snakemake@params$test_years
 
 # test
-# filein <- "results/data/climate/guyaflux.tsv"
-# folderin <- "results/simulations/current/10-years/guyaflux/warmup/R1"
-# folderout <- "results/simulations/current/10-years/guyaflux/historical/R1"
-# figout <- "results/simulations/current/10-years/guyaflux/historical/R1.png"
-# type <- "current"
-# period <- "10-years"
-# climate <- "guyaflux"
-# exp <- "historical"
+# filein <- "results/data/climate/MPI-M-MPI-ESM_ICTP-RegCM4-7.tsv"
+# folderin <- "results/simulations/projections/full/MPI-M-MPI-ESM_ICTP-RegCM4-7/warmup/R1"
+# type <- "projection"
+# period <- "full"
+# climate <- "MPI-M-MPI-ESM_ICTP-RegCM4-7"
+# exp <- "rcp85"
 # rep <- "R1"
 # verbose <- TRUE
 # test <- TRUE
@@ -37,25 +35,29 @@ suppressMessages(library(rcontroll))
 suppressMessages(library(vroom)) 
 
 # code
-path <- gsub(rep, "", folderout)
+# path <- gsub(rep, "", folderout)
 # dir.create(path,
 #            recursive = TRUE)
 
 # climate
 data <- vroom::vroom(filein)
+e <- exp
 if("exp" %in% names(data)) # for cordex
-  data <- filter(data, exp == "historical") %>% 
-  select(-exp)
+  data <- filter(data, exp == e) %>% 
+  mutate(month = month(time), day = day(time)) %>% 
+  mutate(monthday = paste0(month, "-", day)) %>% 
+  filter(monthday != "2-29") %>% 
+  select(-exp, -month, -day, -monthday)
 
 clim <- generate_climate(data)
 day <- generate_dailyvar(data)
 
 # other data
-n <- as.numeric(nrow(clim))
+n <- as.numeric(nrow(clim))*2
 if(test)
   n <- round(test_years*365)
 global_pars <- vroom(file.path(folderin, paste0(rep, "_input_global.txt"))) %>% 
-  mutate(value, ifelse(param == "nbiter", n, value))
+  mutate(value = ifelse(param == "nbiter", n, value))
 species_pars <- vroom(file.path(folderin, paste0(rep, "_input_species.txt")))
 pedo_pars <- vroom(file.path(folderin, paste0(rep, "_input_pedology.txt")))
 forest_pars <- vroom(file.path(folderin, paste0(rep, "_0_final_pattern.txt"))) %>%